refactor: deleted unnecessary files & styles. moved form files in one folder
